refactor ApiCalls file
