+++ src/api/course/category.js
@@ -1,0 +1,10 @@
+import api from "../ApiUrl"
+
+export const getCategories = async () => {
+  try {
+    const response = await api.get('course_categories/')
+    return response
+  } catch (error) {
+    return error
+  }
+}
