+++ src/api/course/index.js
@@ -1,0 +1,6 @@
+export * from './course'
+export * from './category'
+export * from './module'
+export * from './subModule'
+export * from './lesson'
+export * from './quiz'
