+++ src/api/course/quiz.js
@@ -1,0 +1,22 @@
+import api from "../ApiUrl"
+
+export const getQuiz = async id => {
+  try {
+    const response = await api.get(`quiz/${id}`)
+    return response
+  } catch (error) {
+    return error
+  }
+}
+
+export const addQuestion = async (quizId, data) => {
+  try {
+    const response = await api.post(`mcq_question/create`, {
+      question: data,
+      quiz: quizId,
+    })
+    return response
+  } catch (error) {
+    return error
+  }
+}
